feat!: rename colour packs to themes, reserving skin for shapes
A theme is colours; a skin will be geometry. The four built-ins are purely
colour - bars, grid, peak, scope, darken - so calling them themes describes
what they already are.

The OSC 4 reader becomes Palette, which is what it holds: sixteen colour
slots a terminal reports. That frees Theme for the file format, and makes
the sentence in its own docs true - a theme that names green is asking the
palette what green is.

BREAKING CHANGE: --skin is now --theme, and skins/*.toml is themes/*.toml.
